Update schema and refactor code for dynamic table reference
The schema has been updated to support new field types and to allow parameters to be dynamic. The code has been refactored accordingly to adapt to the new schema, particularly to permit selection of specific tables, rather than referring to a singular, static one. As part of this change, "document" has been replaced with a more generic term "json", providing more flexibility and less ambiguity.